Fix variable count in outlierPCA
outlierPCA oddly appeared to use the number of replicates rather than the number of variables to determine how many dimensions it was appropriate to request. It also implicitly assumed that all samples in a dataset had the same number of replicates and used the number of replicates for the first sample.
It's a simple fix: the length of the "varnames" entry in the input dataset is the number of variables, which is a constant in the dataset and does not relate the number of replicates for each sample.